Scheduler Code Deep Cleanup Phase 1: Move the main Schedule() function in generic_scheduler.go to scheduler.go
Jira: ENTOS-2155
